php*-vld in github --HEAD (+1 squashed commit) Squashed commits: [a261a19] php*-vld: Logic Dumper hooks into the Zend Engine and dumps all the opcodes
The Vulcan Logic Dumper hooks into the Zend Engine and dumps all the
opcodes (execution units) of a script. It can be used to see what is
going on in the Zend Engine.

This is an extension for advanced users and can not be used to
un-encode PHP code that has been encoded with any encoder. If you can't
make it work then this tool is not for you. If you are certain there is
a bug, then please provide a patch (or a good bug report).