Ability to install php-fpm and apache 2 handler at same time
This change allows the ability to override the disabling of `libphp5.so`
by the `--with-fpm` and `--with-cgi` flags.

A new flag is introduced `--with-apache` which will force the build
of the apache handler even in the presence of `--with-fpm`.

To install `libphp5.so` and php-fpm:

`brew install php5X --with-fpm --with-apache`