phpmyadmin: delete raise in class body
This caused phpmyadmin to raise *any* time the formula was being
loaded, regardless of whether it was being installed. For instance,
in brew doctor.

This was also unnecessary, since the --without-mcrypt option was
conditional to begin with. The option didn't even exist on the machines
where this conditional was coming up.

Closes #1098.