Provide a high level composer method
The `composer` method runs composer with the `--no-interaction` flag
because the user is not able to interact with it during a brew run.

We also add `--verbose` to composer if brew is run with `--verbose`.

We also provide a method for ensuring the Github API token is configured
if we know one.

A method is also provided for running `composer install` with all of the
above taken care of.