php70-gearman 2.0.1 (new formula)
Adds a new php70-gearman formula that updates php56-gearman and uses wcgallego/pecl-gearman.

This is because PECL doesn't yet contain a Gearman extension for PHP 7; all efforts to this
end are being made by the community - which is fine, but makes life difficult for those of us
who want to use PHP 7 in the meantime.

Closes #3497.

Signed-off-by: Jan Viljanen <jan.viljanen@greenpeace.org>