Be more strict about conflicting PHP formulae
At the moment, any formula ending with `php\d\d` will conflict with a php formula upgrade.
This is wrong as some third party provides PHP extensions compatibles with homebrew-php system and could be broken because of this. `blackfire-php70` recipe for instance conflicts on upgrade whereas it should not.

This patch fixes this issue by being more strict about "our self"

Closes #3245.

Signed-off-by: Andy Blyler <andy@blyler.cc>